% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{predict.bestSubset}
\alias{predict.bestSubset}
\title{Predict method for bestSubset objects}
\usage{
\method{predict}{bestSubset}(object, newdata, type = c("class", "prob"), ...)
}
\arguments{
\item{object}{A bestSubset object from bestSubset()}

\item{newdata}{A matrix or data frame of new predictor values}

\item{type}{Type of prediction: "class" for binary predictions or "prob" for probabilities}

\item{...}{Additional arguments (currently unused)}
}
\value{
A numeric vector of predictions
}
\description{
Predict method for bestSubset objects
}
